implement job queue class
